# AppBackend/openapi.yaml
openapi: 3.0.3
info:
  title: PhotoApp Backend API
  version: 1.0.0
  description: OpenAPI specification for PhotoApp backend (auto-generated from controller files)
servers:
  - url: http://localhost:3000
...
paths:
...
  /albums:
    post:
      tags:
        - Albums
      summary: Create album (auth required)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                hexcolor:
                  type: string
                year:
                  type: integer
                coverimage:
                  type: string
                categoryId:
                  type: integer
      responses:
        '201':
          description: Album created
    get:
      tags:
        - Albums
      summary: Get user albums with categories (auth required)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Albums list
  /albums/simple:
    get:
      tags:
        - Albums
      summary: Get simple albums (auth required)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Albums list
  /albums/{id}:
    get:
      tags:
        - Albums
      summary: Get specific album by id (auth required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Album
    put:
      tags:
        - Albums
      summary: Update album (auth required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                hexcolor:
                  type: string
                year:
                  type: integer
                coverimage:
                  type: string
                categoryId:
                  type: integer
      responses:
        '200':
          description: Updated album
  /albums/{id}/delete:
    post:
      tags:
        - Albums
      summary: Delete album (auth required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
...
  /albums/{id}/photos:
    post:
      tags:
        - Albums
      summary: Add photo to album (auth required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                photoName:
                  type: string
                photoUrl:
                  type: string
      responses:
        '200':
          description: Photo added
    get:
      tags:
        - Albums
      summary: Get photos from album (auth required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Album photos
  /albums/{id}/photos/batch:
    post:
      tags:
        - Albums
      summary: Batch add photos to album (auth required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                photos:
                  type: array
                  items:
                    type: object
                    properties:
                      photoName:
                        type: string
                      photoUrl:
                        type: string
      responses:
        '200':
          description: Batch add result
  /albums/{id}/photos/{photoName}/remove:
    post:
      tags:
        - Albums
      summary: Remove photo from album (auth required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: photoName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Photo removed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT